requirements.txt
openai>=1.0.0
python-docx>=1.0.0
chromadb>=0.4.0
streamlit>=1.28.0
python-dotenv>=1.0.0
